Configuração das validações do fomurlário de criação de agendamentos e ajustes no base.html

+++ procedimentos/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Procedimento(models.Model):
     procedimento = models.TextField()
-    valor = models.IntegerField()
+    valor = models.DecimalField(max_digits=10, decimal_places=2)
     
     def __str__(self):
         return self.procedimento
